embedding-service/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic

pdfplumber
openai
qdrant-client
